+++ src/traintestgyms/ignitegym.py
@@ -65,7 +65,7 @@ class TrainTestGym:
 
     ### Training step
     def __train_step(self, engine, batch):
-        batch = batch.to(self.device, dtype=torch.float16)
+        batch = batch.to(self.device)
 
         if engine.t_start != 0:
             engine.t_start = batch[0,self.time_column_idx].item()
